services/segmentation/requirements.txt
torch
torchvision
transformers
Pillow
numpy
flask
gunicorn
matplotlib